refactor: Fix grammar, separate data source based on function
