+++ shared/src/commonMain/kotlin/id/nisyafawwaz/nyampur/data/repository/AuthRepositoryImpl.kt
@@ -1,20 +1,20 @@
 package id.nisyafawwaz.nyampur.data.repository
 
-import id.nisyafawwaz.nyampur.data.remote.datasources.SupabaseDataSource
+import id.nisyafawwaz.nyampur.data.remote.datasources.supabase.SupabaseAuthDataSource
 import id.nisyafawwaz.nyampur.domain.repository.AuthRepository
 import io.github.jan.supabase.gotrue.user.UserInfo
 
-class AuthRepositoryImpl(private val supabaseDataSource: SupabaseDataSource) : AuthRepository {
+class AuthRepositoryImpl(private val supabaseRecipeDataSource: SupabaseAuthDataSource) : AuthRepository {
 
     override suspend fun sendEmailSignInOtp(email: String) {
-        supabaseDataSource.sendEmailSignInOtp(email)
+        supabaseRecipeDataSource.sendEmailSignInOtp(email)
     }
 
     override suspend fun validateEmailOtp(token: String, email: String) {
-        supabaseDataSource.validateEmailOtp(token, email)
+        supabaseRecipeDataSource.validateEmailOtp(token, email)
     }
 
     override suspend fun retrieveUserSession(): UserInfo? {
-        return supabaseDataSource.retrieveUserSession()
+        return supabaseRecipeDataSource.retrieveUserSession()
     }
 }
