+++ shared/src/commonMain/kotlin/id/nisyafawwaz/nyampur/data/repository/RecipeRepositoryImpl.kt
@@ -2,14 +2,14 @@ package id.nisyafawwaz.nyampur.data.repository
 
 import id.nisyafawwaz.nyampur.data.models.responses.RecipeResponse
 import id.nisyafawwaz.nyampur.data.remote.datasources.RemoteDataSource
-import id.nisyafawwaz.nyampur.data.remote.datasources.SupabaseDataSource
+import id.nisyafawwaz.nyampur.data.remote.datasources.supabase.SupabaseRecipeDataSource
 import id.nisyafawwaz.nyampur.domain.models.RecipeModel
 import id.nisyafawwaz.nyampur.domain.repository.RecipeRepository
 import id.nisyafawwaz.nyampur.utils.enums.SortType
 
 class RecipeRepositoryImpl(
     private val remoteDataSource: RemoteDataSource,
-    private val supabaseDataSource: SupabaseDataSource
+    private val supabaseRecipeDataSource: SupabaseRecipeDataSource
 ) : RecipeRepository {
 
     override suspend fun getRecipe(type: String, page: Int): List<RecipeModel> {
@@ -17,14 +17,14 @@ class RecipeRepositoryImpl(
     }
 
     override suspend fun getSavedRecipes(usersId: String, sortType: SortType): List<RecipeModel> {
-        return supabaseDataSource.getSavedRecipes(usersId, sortType).map(RecipeModel::from)
+        return supabaseRecipeDataSource.getSavedRecipes(usersId, sortType).map(RecipeModel::from)
     }
 
     override suspend fun saveRecipe(response: RecipeResponse): RecipeModel {
-        return RecipeModel.from(supabaseDataSource.saveRecipe(response))
+        return RecipeModel.from(supabaseRecipeDataSource.saveRecipe(response))
     }
 
     override suspend fun deleteSavedRecipe(response: RecipeResponse): RecipeModel {
-        return RecipeModel.from(supabaseDataSource.deleteSavedRecipe(response))
+        return RecipeModel.from(supabaseRecipeDataSource.deleteSavedRecipe(response))
     }
 }
