+++ shared/src/commonMain/kotlin/id/nisyafawwaz/nyampur/di/DataSourceModule.kt
@@ -1,11 +1,13 @@
 package id.nisyafawwaz.nyampur.di
 
 import id.nisyafawwaz.nyampur.data.remote.datasources.RemoteDataSource
-import id.nisyafawwaz.nyampur.data.remote.datasources.SupabaseDataSource
+import id.nisyafawwaz.nyampur.data.remote.datasources.supabase.SupabaseAuthDataSource
+import id.nisyafawwaz.nyampur.data.remote.datasources.supabase.SupabaseRecipeDataSource
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
 
 val dataSourceModule = module {
     singleOf(::RemoteDataSource)
-    singleOf(::SupabaseDataSource)
+    singleOf(::SupabaseRecipeDataSource)
+    singleOf(::SupabaseAuthDataSource)
 }
